further munging. nuked fake_couchdb in advance of moving to mox for everything
